backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
pyodbc==5.1.0
PyJWT==2.8.0
python-dotenv==1.0.1
bcrypt==4.1.3
Werkzeug==3.0.4
